pyproject.toml
[project]
name = "multi-agent-codenames"
version = "2.0.0"
description = "Multi-Agent Codenames — CrewAI + Gemini powered board game"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "crewai>=0.100.0",
    "crewai-tools>=0.17.0",
    "tavily-python>=0.5.0",
    "fastapi>=0.115.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "uvicorn[standard]>=0.32.0",
    "websockets>=13.0",
    "colorama>=0.4.0",
    "pyyaml>=6.0",
]

[project.scripts]
codenames = "main:main"
